refactor: file structures refactoring
