--- xmart-frontend/src/main/resources/primary.fxml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.layout.VBox?>
-
-<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.ezip.ing1.pds.PrimaryController">
-    <children>
-        <Label text="Primary View" />
-        <Button fx:id="primaryButton" onAction="#switchToSecondary" text="Switch to Secondary View" />
-        <TextArea prefHeight="200.0" prefWidth="200.0" />
-    </children>
-    <padding>
-        <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
-    </padding>
-</VBox>
