+++ xmart-frontend/src/main/resources/views/mainView.fxml
